// src/interfaces/purchase/purchase.ts
import { ICurrency } from "../currency/currency";
import { ISupplier } from "../supplier/supplier";
import { IDocumentType } from "../document-type/documentType";
import { IEstablishment } from "../establishment/establishment";
import { IPaymentTerm } from "../payment-term/paymentTerm";
import { IUser } from "../user/user";
import { IWarehouse } from "../warehouse/warehouse";
import { IPurchaseInstalment } from "./instalment";
import { IPurchasePayment } from "./payment";
import { IPurchaseProduct } from "./product";

export interface IPurchase {
    id: number;
    supplier_id: number;
    currency_id: number;
    establishment_id: number;
    warehouse_id: number;
    document_type_id: number;
    payment_term_id: number;
    amount: number;
    paid_amount: number;
    purchased_at: string;
    expires_at: string | null;
    paid_at: string | null;
    name: string;
    identification_document: string;
    phone: string | null;
    email: string | null;
    address: string | null;
    comments: string | null;
    created_by: number | null;
    updated_by: number | null;
    created_at: string;
    updated_at: string;
    supplier: ISupplier;
    currency: ICurrency;
    establishment: IEstablishment;
    warehouse: IWarehouse;
    document_type: IDocumentType;
    payment_term: IPaymentTerm;
    products: IPurchaseProduct[];
    payments: IPurchasePayment[];
    instalments: IPurchaseInstalment[];
};